telegram_bot/requirements.txt
python-dotenv==1.1.0
aiogram==3.20.0.post0
structlog==25.3.0
asyncpg==0.30.0
aiocache==0.12.3
openpyxl==3.1.5